+++ ChangeLog.md
@@ -7,6 +7,7 @@ Note that 0.0.8.2 has been skipped, it was an experimental build distributed to 
 - Added Orion and Orion alpha armor pieces
 - Fixed Kulve Taroth armor pieces defense based on user report
 - Fixed issue related to crashes when loading of System.Net.Http version 4.2.0.0
+- Fixed wrong gender processing for armor pieces selection
 
 ## 0.0.8.1 (2018/12/13)